third_party/blink/web_tests/fast/writing-mode/percentage-padding-expected.txt

PASS: The block is the correct size.
